read cookie from request
