+++ runtimes/core/src/api/schema/header.rs
@@ -1,5 +1,3 @@
-use http::header::COOKIE;
-
 use crate::api::reqauth::meta::HeaderValueExt;
 use crate::api::schema::{JSONPayload, ParseResponse, ToOutgoingRequest, ToResponse};
 use crate::api::{self, PValue, PValues};
@@ -58,38 +56,6 @@ impl Header {
             })
     }
 }
-
-#[derive(Debug, Clone)]
-pub struct Cookie {
-    schema: jsonschema::JSONSchema,
-}
-
-impl Cookie {
-    pub fn new(schema: jsonschema::JSONSchema) -> Self {
-        Self { schema }
-    }
-
-    pub fn parse(&self, headers: &axum::http::HeaderMap) -> APIResult<Option<PValues>> {
-        if self.schema.root().is_empty() {
-            return Ok(None);
-        }
-
-        let mut jar = cookie::CookieJar::new();
-        for raw in headers.get_all(COOKIE) {
-            for raw_part in String::from_utf8_lossy(raw.as_bytes()).split(',') {
-                let c = cookie::Cookie::parse(raw_part).map_err(api::Error::internal)?;
-                jar.add_original(c.into_owned());
-            }
-        }
-
-        match self.schema.parse(jar) {
-            Ok(decoded) => Ok(Some(decoded)),
-            Err(err) => Err(err),
-        }
-    }
-}
-
-pub struct RawCookie {}
 
 pub trait AsStr {
     fn as_str(&self) -> &str;
